backend/requirements.txt
python-dotenv
pymongo
newspaper3k
requests
uuid
